Improving reset username condition #UL-IE-1055
The previous criteria were not 100% correct. After discussing this issue
with Daniela Mendes we decided to change the conditions a bit, our main
focus is to reset pending accounts from their temporary CN back to
fenix's bennu username.

To do this we have to check for 3 things:

1. User's CN is the identification document: this means the user is in
pending state at the IDM
2. User's username at fenix is bennuXXXX: this means the user never
logged-in in fenix
3. Student, Employee and Teacher flags are all false: this means it's
not an active user in the system.

Keep in mind that we do check for the Alumni flag, in the Alumni case we
do have 2 cases:

1. The user has a valid campus account
2. The user never had a campus account

We will only want to reset the 2nd case which is already covered by the
condition 1, because if has no campus account then it will be in pending
state.