Changing personal email collection strategy
When a person had several personal emails we would use the first one we
found. Now we first try to detect if there's one that is marked as
default if so, that one is used. Otherwise we fallback to the previous
strategy where we use the first we find.

Relates with: #UL-FD-1978